Prepare the expression data by selecting basal-like/classical genes and running normalization

The function selects the expression values related to the gene list passed in argument. Then, the function adds one to the expression values before log transforming the values. Then, the function normalizes the data so that each row has mean 0 and variance 1.

prepareExpression(rna_data, gene_list, gene_column = "GENE", log_data = TRUE)

Arguments

rna_data

a data.frame containing the expression values for the studied samples. There must be one column per sample and an extra column for the gene names. The name of the column containing the gene names is passed in the gene_column parameter.

gene_list

a data.frame containing the gene list used for the basal-like/classical classification.

gene_column

a character string representing the name of the column in the rna_data parameter that contains the gene names.

log_data

a logical indicating if the data should be log transformed. If so, the data will be log2(value + 1) transformed. Default: TRUE.

Value

a list containing 2 entries:

DATA a data.frame containing the normalized expression values for the genes that are part of the classical/basal-like gene list. The gene names are used as row names and the gene order fits the order of the gene names in the ROW_INFO data.frame.

ROW_INFO a data.frame containing information about the genes from basal-like/classical classification that are present in the DATA data.frame. The gene names are used as row names and the gene order fits the order of the gene names in the DATA data.frame.

Author

Astrid Deschênes

Examples

## Load Moffitt et al 2015 gene list
data("Moffitt_2015")

## Load demo PDAC TCGA expression data
data("demo_PDAC_TCGA")

## Prepare the data for the heatmap creation step
result <- SignatureHeatmap:::prepareExpression(rna_data=demo_PDAC_TCGA, 
    gene_list=Moffitt_2015, gene_column="GENE", log_data=TRUE)